requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
python-dotenv==1.1.0
pydantic==2.11.3
pydantic-settings==2.8.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
python-multipart==0.0.20
alembic==1.15.2
email-validator==2.2.0
bcrypt==4.3.0
cryptography==44.0.2